Move Coprimitive to a separate module
