+++ monad-bayes.cabal
@@ -28,6 +28,7 @@ library
                        Control.Monad.Bayes.Particle,
                        Control.Monad.Bayes.Inference,
                        Control.Monad.Bayes.Metrics,
+                       Control.Monad.Bayes.Coprimitive,
                        Control.Monad.Bayes.Trace
   build-depends:       base >= 4.7 && < 5
                      , containers
